backend/supermarket/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.11
django-cors-headers==4.3.0
dj-database-url==1.2.0
pillow==12.0.0
dj-database-url==1.2.0
